backend/requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-gis==0.18
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
mysqlclient==2.1.0
node==1.0
odict==1.8.1
openapi-codec==1.3.2
packaging==21.3
plumber==1.7
PyJWT==2.3.0
pyparsing==3.0.7
python-decouple==3.6
pytz==2018.4
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
zope.proxy==4.5.0
